Add new variables that translate updates to base SORTIE parameter file

treelistDfn() adds additional initial tree diameter size classes and prefixes to VariableNames

Usage

treelistDfn(initname, numDigits = 0, diamMin, diamMax, diamInc)

Arguments

initname

Naming convention character() used to describe diameter size classes

numDigits

Number of decimal places required to match the base parameter file

diamMin

Minimum diameter size

diamMax

Maximum diameter size

diamInc

Size (in cm) of diameter bins

Value

Details

VariableNames is a table that translates the names of parameters found within behaviours (variables) defined by a user in the newvals object to the names of these parameters (variables) found in the base SORTIE parameter file. This file is essential to finding the right variable within the right behaviour to update with new values during a makeFiles() call.

For details on how to write a new VariableNames file, see the Preparing inputs for rsortie vignette, linked below

There is a default VariableNames loaded with the rsortie package, but a user may wish to add additional variable translations.

See also

Examples

samplebasexml[1:30]
#>  [1] "<?xml version=\"1.0\" encoding=\"UTF-8\" standalone=\"no\"?>"            
#>  [2] "<paramFile fileCode=\"07050101\">"                                       
#>  [3] "  <plot>"                                                                
#>  [4] "    <timesteps>10</timesteps>"                                           
#>  [5] "    <rt_timestep>0</rt_timestep>"                                        
#>  [6] "    <randomSeed>0</randomSeed>"                                          
#>  [7] "    <yearsPerTimestep>5.0</yearsPerTimestep>"                            
#>  [8] "    <plot_lenX>300.0</plot_lenX>"                                        
#>  [9] "    <plot_lenY>300.0</plot_lenY>"                                        
#> [10] "    <plot_latitude>41.92</plot_latitude>"                                
#> [11] "    <plot_precip_mm_yr>0.0</plot_precip_mm_yr>"                          
#> [12] "    <plot_seasonal_precipitation>0.0</plot_seasonal_precipitation>"      
#> [13] "    <plot_water_deficit>0.0</plot_water_deficit>"                        
#> [14] "    <plot_temp_C>0.0</plot_temp_C>"                                      
#> [15] "    <plot_n_dep>0.0</plot_n_dep>"                                        
#> [16] "    <plot_title>Default parameter file-use for testing only</plot_title>"
#> [17] "  </plot>"                                                               
#> [18] "  <trees>"                                                               
#> [19] "    <tr_speciesList>"                                                    
#> [20] "      <tr_species speciesName=\"ACRU\"/>"                                
#> [21] "      <tr_species speciesName=\"ACSA\"/>"                                
#> [22] "      <tr_species speciesName=\"BEAL\"/>"                                
#> [23] "      <tr_species speciesName=\"FAGR\"/>"                                
#> [24] "      <tr_species speciesName=\"TSCA\"/>"                                
#> [25] "      <tr_species speciesName=\"FRAM\"/>"                                
#> [26] "      <tr_species speciesName=\"PIST\"/>"                                
#> [27] "      <tr_species speciesName=\"PRSE\"/>"                                
#> [28] "      <tr_species speciesName=\"QURU\"/>"                                
#> [29] "    </tr_speciesList>"                                                   
#> [30] "    <tr_sizeClasses>"                                                    
VariableNames[1:30,]
#>         InputFileParameterName Type                   code.name
#> 1                    Timesteps    1                   timesteps
#> 2                   Randomseed    1                  randomSeed
#> 3             YearsPerTimestep    1            yearsPerTimestep
#> 4                  rt_timestep    1                 rt_timestep
#> 5                   randomSeed    1                  randomSeed
#> 6             yearsPerTimestep    1            yearsPerTimestep
#> 7                    plot_lenX    1                   plot_lenX
#> 8                    plot_lenY    1                   plot_lenY
#> 9                plot_latitude    1               plot_latitude
#> 10           plot_precip_mm_yr    1           plot_precip_mm_yr
#> 11 plot_seasonal_precipitation    1 plot_seasonal_precipitation
#> 12          plot_water_deficit    1          plot_water_deficit
#> 13                 plot_temp_C    1                 plot_temp_C
#> 14                  plot_n_dep    1                  plot_n_dep
#> 15                  plot_title    1                         sph
#> 16                        !nci                                 
#> 17       nciWhichShadingEffect    3       nciWhichShadingEffect
#> 18      nciWhichCrowdingEffect    3      nciWhichCrowdingEffect
#> 19             nciWhichNCITerm    3             nciWhichNCITerm
#> 20          nciWhichSizeEffect    3          nciWhichSizeEffect
#> 21        nciWhichDamageEffect    3        nciWhichDamageEffect
#> 22 nciWhichPrecipitationEffect    3 nciWhichPrecipitationEffect
#> 23   nciWhichTemperatureEffect    3   nciWhichTemperatureEffect
#> 24      nciWhichNitrogenEffect    3      nciWhichNitrogenEffect
#> 25     nciWhichInfectionEffect    3     nciWhichInfectionEffect
#> 26               NCI.MaxGrowth    4          gr_nmpgVal species
#> 27        Std.Dev.Norm.Log.Adj    4            gr_sdVal species
#> 28         NCI.Shading.Coeff.m    4              nscVal species
#> 29           NCI.Shading.Exp.n    4              nseVal species
#> 30             NCI.Size.Sensit    4              ncgVal species
#>                  group.name
#> 1                          
#> 2                          
#> 3                          
#> 4                          
#> 5                          
#> 6                          
#> 7                          
#> 8                          
#> 9                          
#> 10                         
#> 11                         
#> 12                         
#> 13                         
#> 14                         
#> 15               plot_title
#> 16                         
#> 17                         
#> 18                         
#> 19                         
#> 20                         
#> 21                         
#> 22                         
#> 23                         
#> 24                         
#> 25                         
#> 26 gr_nciMaxPotentialGrowth
#> 27     gr_standardDeviation
#> 28    nciShadingCoefficient
#> 29       nciShadingExponent
#> 30         nciCrowdingGamma